Added the essential implementation of the core operations of the library.
The API has three C functions, which are meant to be executed by the
memory segment creator and the segment accessor processes. A driver to
act as the segment creator shows the main functionality. Once executed,
the code creates the memory segment and then detaches from it prior to
terminating. Any subsequent calls to "ipcs" from the shell will not show
the memory segment that was created as being in existence, indicating
that it was cleaned-up. Modifying this code with waiting statements, like
sleep(), will allow querying the existence of the memory segment from
other shells.

IN 2018/05/27