Begun authoring Fortran interface of the library. This work is incomplete.
At present, shared memory segments in Fortran are managed using handles,
but they are not directly usable as they do not attach their memory to
specific pointers on the fortran side. This is to be completed and tested
using the "demo2" for which a placeholder exists in the Makefile.
IN 2018/05/27